Organized Header components inside section folder. Added toggle active class to menubutton.
