+++ src/components/About.jsx
@@ -1,0 +1,7 @@
+const About = () => {
+  return (
+    <div>About</div>
+  )
+}
+
+export { About }
