+++ src/components/Header/LangButton.jsx
@@ -1,5 +1,5 @@
-import esFlag from '../assets/es-flag.svg'
-import enFlag from '../assets/en-flag.svg'
+import esFlag from '../../assets/es-flag.svg'
+import enFlag from '../../assets/en-flag.svg'
 
 const LangButton = (props) => {
     const handleChange = (evt) => {
