+++ src/components/Header/Logo.jsx
@@ -1,4 +1,4 @@
-import brandLogo from '../assets/logo.png'
+import brandLogo from '../../assets/logo.png'
 
 const Logo = () => {
   return (
