+++ src/styles/Footer.component.css
@@ -47,5 +47,9 @@
 
 .Footer__topBtn a {
   color: var(--white-color);
-  font-size: 2.8rem;
+  font-size: 3.2rem;
+  width: 100%;
+  height: 100%;
+  padding: 7px 0;
+  text-align: center;
 }
