+++ src/styles/Header.component.css
@@ -34,8 +34,8 @@
 }
 
 .Logo img {
-  width: 56px;
-  height: 56px;
+  width: 48px;
+  height: 48px;
 }
 
 .Logo span {
