+++ src/styles/Hero.component.css
@@ -4,7 +4,7 @@
     min-width: 320px;
     min-height: 587px;
     margin-top: var(--top-margin);
-    background: url(../assets/code_cover.jpg) no-repeat;
+    background: url(../assets/cover_code.jpg) no-repeat;
     background-size: cover;
     display: flex;
     flex-direction: column;
@@ -18,7 +18,7 @@
     border-radius: 15px;
     margin-bottom: 80px;
     background-color: var(--dark-color-1);
-    box-shadow: 1px 1px 4px 5px rgba(0, 0, 0, 0.35);
+    box-shadow: 1px 1px 4px 5px rgba(0, 0, 0, 0.25);
     color: var(--white-color);
     display: flex;
     flex-direction: column;
